Add virsh provider [1]
Cleanup
