+++ tests/test-stream.yaml
@@ -1,0 +1,2 @@
+{'type': 'ref-updated', 'refUpdate': {'newRev': '1a8c2369ed141a50ac5c109c53cf5fde68375afc', 'refName': 'master', 'oldRev': '0fc633318d0103e81af00840679661d3df0534aa', 'project': 'stackforge/stackalytics'}}
+{"type":"comment-added","change":{"project":"openstack-dev/ci-sandbox","branch":"master","id":"I8f4016656ea4adaad382263eb6f223d1f6e7876a","number":"189221","subject":"test","owner":{"name":"zhaohua","email":"zhaohua.zhao@huawei.com","username":"zhaohua"},"url":"https://review.openstack.org/189221","commitMessage":"test\n\nChange-Id: I8f4016656ea4adaad382263eb6f223d1f6e7876a\n","status":"NEW"},"patchSet":{"number":"1","revision":"1ab211f25fb4df285a6f726212bccaa9417fe96c","parents":["8e42d6543eb15ad2a415171b93185a078b7baace"],"ref":"refs/changes/21/189221/1","uploader":{"name":"zhaohua","email":"zhaohua.zhao@huawei.com","username":"zhaohua"},"createdOn":1433752243,"author":{"name":"zhaohua","email":"zhaohua.zhao@huawei.com","username":"zhaohua"},"isDraft":false,"sizeInsertions":0,"sizeDeletions":0},"author":{"name":"Sergey Skripnick","email":"sskripnick@mirantis.com","username":"redixin"},"comment":"Patch Set 1:\n\nmy-ci recheck"}
